refactor: consolidate objective/observable hooks onto BaseEvaluator
compute_objective, compute_objective_from_shots, compute_observables, and
the outputs-merging tail were duplicated verbatim between
MultiDeviceScanEvaluator and ScalarLogEvaluator. Moved all four to
BaseEvaluator as a shared _compute_outputs helper; both _get_value
implementations now build their shot-list and delegate.

Also removed the dead evaluation_mode field (stored but never read) from
BaseOptimizer and BaseOptimizerConfig, and stripped the leftover
TYPE_CHECKING/pass block from bcavemagspec_opt.py.
