+++ GEECS-Scanner-GUI/geecs_scanner/optimization/evaluators/multi_device_scan_evaluator.py
@@ -26,9 +26,7 @@ from __future__ import annotations
 
 import importlib
 import logging
-from typing import TYPE_CHECKING, Dict, List, Optional, Union
-
-import numpy as np
+from typing import TYPE_CHECKING, Dict, List, Optional
 
 if TYPE_CHECKING:
     from geecs_scanner.data_acquisition.data_logger import DataLogger
@@ -231,121 +229,4 @@ class MultiDeviceScanEvaluator(BaseEvaluator):
                             device_name,
                         )
 
-        outputs: Dict[str, float] = {}
-        output_key = self.output_key
-
-        if output_key is not None:
-            objective_result = self.compute_objective_from_shots(
-                scalar_results_list=merged, bin_number=self.bin_number
-            )
-            if isinstance(objective_result, dict):
-                if output_key not in objective_result:
-                    logger.warning(
-                        "compute_objective_from_shots dict missing key '%s'", output_key
-                    )
-                outputs.update({str(k): float(v) for k, v in objective_result.items()})
-            else:
-                outputs[output_key] = float(objective_result)
-
-        # mean-aggregate merged slots for compute_observables
-        all_keys = set().union(*(d.keys() for d in merged)) if merged else set()
-        aggregated: Dict[str, float] = {
-            k: float(np.mean([d[k] for d in merged if k in d])) for k in all_keys
-        }
-
-        extra = (
-            self.compute_observables(
-                scalar_results=aggregated, bin_number=self.bin_number
-            )
-            or {}
-        )
-        if output_key is not None and output_key in extra:
-            logger.warning(
-                "compute_observables returned objective key '%s'; removing", output_key
-            )
-            extra = {k: v for k, v in extra.items() if k != output_key}
-
-        for k, v in extra.items():
-            if str(k) not in outputs:
-                outputs[str(k)] = float(v)
-
-        return outputs
-
-    # ------------------------------------------------------------------
-    # Hooks for subclasses
-    # ------------------------------------------------------------------
-
-    def compute_objective(self, scalar_results: dict, bin_number: int) -> float:
-        """
-        Compute the objective from aggregated scalar results.
-
-        Override this for simple evaluators where mean aggregation across shots
-        is sufficient.  For full per-shot control (median, noise estimate, etc.)
-        override :meth:`compute_objective_from_shots` instead.
-
-        Raises
-        ------
-        NotImplementedError
-            When neither this method nor :meth:`compute_objective_from_shots`
-            is overridden.
-        """
-        raise NotImplementedError(
-            f"{self.__class__.__name__} must implement compute_objective "
-            "or override compute_objective_from_shots."
-        )
-
-    def compute_objective_from_shots(
-        self,
-        scalar_results_list: List[Dict[str, float]],
-        bin_number: int,
-    ) -> Union[float, Dict[str, float]]:
-        """
-        Compute the objective from per-slot scalar results.
-
-        The default mean-aggregates all slots and delegates to
-        :meth:`compute_objective`, so subclasses that implement only
-        ``compute_objective`` work for all analysis modes without changes.
-
-        Override this to apply custom statistics, e.g.::
-
-            def compute_objective_from_shots(self, scalar_results_list, bin_number):
-                vals = [self.get_scalar(self.primary_device, "image_total", r)
-                        for r in scalar_results_list]
-                return {"f": -float(np.median(vals)), "f_noise": float(np.std(vals))}
-
-        Parameters
-        ----------
-        scalar_results_list : list of dict
-            One dict per slot (shot or bin, depending on ``analysis_mode``).
-        bin_number : int
-
-        Returns
-        -------
-        float or dict
-            Scalar objective, or a dict containing at least ``self.output_key``
-            plus any extra keys (e.g. ``f_noise``) to pass through to Xopt.
-        """
-        if not scalar_results_list:
-            logger.warning(
-                "compute_objective_from_shots received empty list for bin %s",
-                bin_number,
-            )
-            return 0.0
-
-        all_keys = set().union(*scalar_results_list)
-        aggregated: Dict[str, float] = {
-            k: float(np.mean([d[k] for d in scalar_results_list if k in d]))
-            for k in all_keys
-        }
-        return self.compute_objective(scalar_results=aggregated, bin_number=bin_number)
-
-    def compute_observables(
-        self, scalar_results: dict, bin_number: int
-    ) -> Dict[str, float]:
-        """
-        Return extra scalar observables to include alongside the objective.
-
-        *scalar_results* is always the mean-aggregated dict across all slots.
-        The default returns an empty dict.
-        """
-        return {}
+        return self._compute_outputs(merged, self.bin_number)
